failed attempt at writing smooth world movement
derp, It has problems with the mouse picking

+++ olcLevelMaker/olcConsoleGameEngine.h
@@ -600,7 +600,7 @@ public:
 
 		// Create translated model vector of coordinate pairs
 		vector<pair<float, float>> vecTransformedCoordinates;
-		size_t verts = vecModelCoordinates.size();
+		int verts = vecModelCoordinates.size();
 		vecTransformedCoordinates.resize(verts);
 
 		// Rotate
